refactor(imports): normalize supabase imports to '@/lib/supabase' in app/api
