+++ app/api/users/tier/route.ts
@@ -3,7 +3,7 @@
 
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
-import { supabaseAdmin } from "../../../../src/lib/supabase";
+import { supabaseAdmin } from "../../../../@/lib/supabase";
 
 export async function GET(req: NextRequest) {
   try {
